+++ test/OpenTelemetry.Tests/Logs/LogRecordTest.cs
@@ -57,6 +57,13 @@ namespace OpenTelemetry.Logs.Tests
             this.logger = this.loggerFactory.CreateLogger<LogRecordTest>();
         }
 
+        private enum Field
+        {
+            FormattedMessage,
+            State,
+            StateValues,
+        }
+
         [Fact]
         public void CheckCateogryNameForLog()
         {
@@ -106,6 +113,119 @@ namespace OpenTelemetry.Logs.Tests
             Assert.Equal(1, state.Count);
 
             Assert.Equal(message.ToString(), state.ToString());
+        }
+
+        [Fact]
+        public void CheckStateCanBeSet()
+        {
+            var message = $"This does not matter.";
+            this.logger.LogInformation(message);
+
+            var logRecord = this.exportedItems[0];
+            logRecord.State = "newState";
+
+            var expectedState = "newState";
+            Assert.Equal(expectedState, logRecord.State);
+        }
+
+        [Fact]
+        public void CheckStateValuesCanBeSet()
+        {
+            try
+            {
+                this.options.ParseStateValues = true;
+
+                this.logger.Log(
+                    LogLevel.Information,
+                    0,
+                    new List<KeyValuePair<string, object>> { new KeyValuePair<string, object>("Key1", "Value1") },
+                    null,
+                    (s, e) => "OpenTelemetry!");
+
+                var logRecord = this.exportedItems[0];
+                var expectedStateValues = new List<KeyValuePair<string, object>> { new KeyValuePair<string, object>("Key2", "Value2") };
+                logRecord.StateValues = expectedStateValues;
+
+                Assert.Equal(expectedStateValues, logRecord.StateValues);
+            }
+            finally
+            {
+                this.options.ParseStateValues = false;
+            }
+        }
+
+        [Fact]
+        public void CheckFormattedMessageCanBeSet()
+        {
+            try
+            {
+                this.options.IncludeFormattedMessage = true;
+                this.logger.LogInformation("OpenTelemetry {Greeting} {Subject}!", "Hello", "World");
+
+                var logRecord = this.exportedItems[0];
+                var expectedFormattedMessage = "OpenTelemetry Good Night!";
+                logRecord.FormattedMessage = expectedFormattedMessage;
+
+                Assert.Equal(expectedFormattedMessage, logRecord.FormattedMessage);
+            }
+            finally
+            {
+                this.options.IncludeFormattedMessage = false;
+            }
+        }
+
+        [Fact]
+        public void CheckStateCanBeSetByProcessor()
+        {
+            using var loggerFactory = LoggerFactory.Create(builder => builder
+                .AddOpenTelemetry(options =>
+                {
+                    options.AddProcessor(new RedactionProcessor(Field.State));
+                    options.AddInMemoryExporter(this.exportedItems);
+                }));
+
+            var logger = loggerFactory.CreateLogger<LogRecordTest>();
+            logger.LogInformation($"This does not matter.");
+
+            var state = this.exportedItems[0].State as IReadOnlyList<KeyValuePair<string, object>>;
+            Assert.Equal("newStateKey", state[0].Key.ToString());
+            Assert.Equal("newStateValue", state[0].Value.ToString());
+        }
+
+        [Fact]
+        public void CheckStateValuesCanBeSetByProcessor()
+        {
+            using var loggerFactory = LoggerFactory.Create(builder => builder
+                .AddOpenTelemetry(options =>
+                {
+                    options.AddProcessor(new RedactionProcessor(Field.StateValues));
+                    options.AddInMemoryExporter(this.exportedItems);
+                    options.ParseStateValues = true;
+                }));
+
+            var logger = loggerFactory.CreateLogger<LogRecordTest>();
+            logger.LogInformation("This does not matter.");
+
+            var stateValue = this.exportedItems[0];
+            Assert.Equal(new KeyValuePair<string, object>("newStateValueKey", "newStateValueValue"), stateValue.StateValues[0]);
+        }
+
+        [Fact]
+        public void CheckFormattedMessageCanBeSetByProcessor()
+        {
+            using var loggerFactory = LoggerFactory.Create(builder => builder
+                .AddOpenTelemetry(options =>
+                {
+                    options.AddProcessor(new RedactionProcessor(Field.FormattedMessage));
+                    options.AddInMemoryExporter(this.exportedItems);
+                    options.IncludeFormattedMessage = true;
+                }));
+
+            var logger = loggerFactory.CreateLogger<LogRecordTest>();
+            logger.LogInformation("OpenTelemetry {Greeting} {Subject}!", "Hello", "World");
+
+            var item = this.exportedItems[0];
+            Assert.Equal("OpenTelemetry Good Night!", item.FormattedMessage);
         }
 
         [Fact]
@@ -680,6 +800,32 @@ namespace OpenTelemetry.Logs.Tests
                 data.BufferLogScopes();
 
                 base.OnEnd(data);
+            }
+        }
+
+        private class RedactionProcessor : BaseProcessor<LogRecord>
+        {
+            private readonly Field fieldToUpdate;
+
+            public RedactionProcessor(Field fieldToUpdate)
+            {
+                this.fieldToUpdate = fieldToUpdate;
+            }
+
+            public override void OnEnd(LogRecord logRecord)
+            {
+                if (this.fieldToUpdate == Field.State)
+                {
+                    logRecord.State = new List<KeyValuePair<string, object>> { new KeyValuePair<string, object>("newStateKey", "newStateValue") };
+                }
+                else if (this.fieldToUpdate == Field.StateValues)
+                {
+                    logRecord.StateValues = new List<KeyValuePair<string, object>> { new KeyValuePair<string, object>("newStateValueKey", "newStateValueValue") };
+                }
+                else
+                {
+                    logRecord.FormattedMessage = "OpenTelemetry Good Night!";
+                }
             }
         }
     }